#365 refactor : Util 의 DateDuration 로직, VO로 객체가 관리하도록 수정
