+++ src/main/java/com/floney/floney/analyze/service/AnalyzeServiceImpl.java
@@ -13,13 +13,12 @@ import com.floney.floney.book.domain.entity.Budget;
 import com.floney.floney.book.domain.entity.Category;
 import com.floney.floney.book.domain.entity.category.BookCategory;
 import com.floney.floney.book.dto.process.AssetInfo;
-import com.floney.floney.book.dto.process.DatesDuration;
 import com.floney.floney.book.repository.BookLineRepository;
 import com.floney.floney.book.repository.BookRepository;
 import com.floney.floney.book.repository.analyze.BudgetRepository;
 import com.floney.floney.book.repository.category.CategoryRepository;
-import com.floney.floney.book.util.DateFactory;
 import com.floney.floney.common.constant.Status;
+import com.floney.floney.common.domain.vo.DateDuration;
 import com.floney.floney.common.exception.book.NotFoundBookException;
 import com.floney.floney.common.exception.book.NotFoundCategoryException;
 import lombok.RequiredArgsConstructor;
@@ -46,8 +45,8 @@ public class AnalyzeServiceImpl implements AnalyzeService {
 
         // 분석 종류 - 지출 or 수입
         Category rootCategory = categoryRepository.findParentCategory(request.getRoot())
-            .orElseThrow(() -> new NotFoundCategoryException(request.getRoot()));
-        DatesDuration duration = DateFactory.getStartAndEndOfMonth(request.getDate());
+                .orElseThrow(() -> new NotFoundCategoryException(request.getRoot()));
+        DateDuration duration = DateDuration.getStartAndEndOfMonth(request.getDate());
         String bookKey = request.getBookKey();
 
         // 부모가 지출 or 수입인 자식 카테고리 조회
@@ -64,12 +63,12 @@ public class AnalyzeServiceImpl implements AnalyzeService {
     @Override
     @Transactional(readOnly = true)
     public AnalyzeResponseByBudget analyzeByBudget(AnalyzeRequestByBudget request) {
-        DatesDuration duration = DateFactory.getStartAndEndOfMonth(request.getDate());
+        DateDuration duration = DateDuration.getStartAndEndOfMonth(request.getDate());
         Book savedBook = findBook(request.getBookKey());
 
         // 자산 조회
         Budget budget = budgetRepository.findBudgetByBookAndDate(savedBook, LocalDate.parse(request.getDate()))
-            .orElse(Budget.init());
+                .orElse(Budget.init());
 
         // 총 수입 조회
         double totalOutcome = bookLineRepository.totalOutcomeMoneyForBudget(request, duration);
@@ -91,7 +90,7 @@ public class AnalyzeServiceImpl implements AnalyzeService {
 
     private Book findBook(String bookKey) {
         return bookRepository.findBookByBookKeyAndStatus(bookKey, Status.ACTIVE)
-            .orElseThrow(() -> new NotFoundBookException(bookKey));
+                .orElseThrow(() -> new NotFoundBookException(bookKey));
     }
 
     private double calculateDifference(AnalyzeByCategoryRequest request, double totalMoney) {
@@ -108,7 +107,7 @@ public class AnalyzeServiceImpl implements AnalyzeService {
 
     private double calculateTotalMoney(List<AnalyzeResponseByCategory> result) {
         return result.stream()
-            .mapToDouble(AnalyzeResponseByCategory::getMoney)
-            .sum();
+                .mapToDouble(AnalyzeResponseByCategory::getMoney)
+                .sum();
     }
 }
