+++ src/main/java/com/floney/floney/book/dto/request/AllOutcomesRequest.java
@@ -1,6 +1,6 @@
 package com.floney.floney.book.dto.request;
 
-import com.floney.floney.book.dto.process.DatesDuration;
+import com.floney.floney.common.domain.vo.DateDuration;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -13,14 +13,14 @@ import java.util.List;
 @NoArgsConstructor
 public class AllOutcomesRequest {
 
-    private  String bookKey;
+    private String bookKey;
 
-    private  List<String> usersEmails;
+    private List<String> usersEmails;
 
     @NotNull(message = "duration을 입력해주세요")
-    private  DatesDuration duration;
+    private DateDuration duration;
 
-    public AllOutcomesRequest(String bookKey, List<String> usersEmails, DatesDuration duration) {
+    public AllOutcomesRequest(String bookKey, List<String> usersEmails, DateDuration duration) {
         this.bookKey = bookKey;
         this.usersEmails = usersEmails;
         this.duration = duration;
