+++ src/main/java/com/floney/floney/book/repository/BookCustomRepositoryImpl.java
@@ -3,9 +3,9 @@ package com.floney.floney.book.repository;
 import com.floney.floney.book.domain.entity.Book;
 import com.floney.floney.book.domain.entity.BookUser;
 import com.floney.floney.book.domain.entity.QBookUser;
-import com.floney.floney.book.dto.process.DatesDuration;
 import com.floney.floney.book.dto.response.BudgetYearResponse;
 import com.floney.floney.book.dto.response.QBudgetYearResponse;
+import com.floney.floney.common.domain.vo.DateDuration;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
@@ -30,16 +30,16 @@ public class BookCustomRepositoryImpl implements BookCustomRepository {
     @Override
     public Optional<Book> findByBookUserEmailAndBookKey(final String userEmail, final String bookKey) {
         final BookUser bookUser = jpaQueryFactory.selectFrom(QBookUser.bookUser)
-            .innerJoin(QBookUser.bookUser.book, book).fetchJoin()
-            .innerJoin(QBookUser.bookUser.user, user).fetchJoin()
-            .where(
-                QBookUser.bookUser.status.eq(ACTIVE),
-                user.email.eq(userEmail),
-                book.bookKey.eq(bookKey),
-                user.status.eq(ACTIVE),
-                book.status.eq(ACTIVE)
-            )
-            .fetchOne();
+                .innerJoin(QBookUser.bookUser.book, book).fetchJoin()
+                .innerJoin(QBookUser.bookUser.user, user).fetchJoin()
+                .where(
+                        QBookUser.bookUser.status.eq(ACTIVE),
+                        user.email.eq(userEmail),
+                        book.bookKey.eq(bookKey),
+                        user.status.eq(ACTIVE),
+                        book.status.eq(ACTIVE)
+                )
+                .fetchOne();
 
         if (bookUser == null) {
             return Optional.empty();
@@ -48,27 +48,27 @@ public class BookCustomRepositoryImpl implements BookCustomRepository {
     }
 
     @Override
-    public List<BudgetYearResponse> findBudgetByYear(String bookKey, DatesDuration duration) {
+    public List<BudgetYearResponse> findBudgetByYear(String bookKey, DateDuration duration) {
         return jpaQueryFactory
-            .select(new QBudgetYearResponse(budget.date, budget.money))
-            .from(budget)
-            .innerJoin(budget.book, book)
-            .where(book.bookKey.eq(bookKey))
-            .where(budget.date.between(duration.getStartDate(), duration.getEndDate()))
-            .fetch();
+                .select(new QBudgetYearResponse(budget.date, budget.money))
+                .from(budget)
+                .innerJoin(budget.book, book)
+                .where(book.bookKey.eq(bookKey))
+                .where(budget.date.between(duration.getStartDate(), duration.getEndDate()))
+                .fetch();
     }
 
     @Override
     public List<Book> findAllByUserEmail(final String userEmail) {
         return jpaQueryFactory.select(book)
-            .from(bookUser)
-            .innerJoin(bookUser.book, book)
-            .innerJoin(bookUser.user, user)
-            .where(
-                user.email.eq(userEmail),
-                book.status.eq(ACTIVE),
-                bookUser.status.eq(ACTIVE)
-            )
-            .fetch();
+                .from(bookUser)
+                .innerJoin(bookUser.book, book)
+                .innerJoin(bookUser.user, user)
+                .where(
+                        user.email.eq(userEmail),
+                        book.status.eq(ACTIVE),
+                        bookUser.status.eq(ACTIVE)
+                )
+                .fetch();
     }
 }
