+++ src/main/java/com/floney/floney/book/repository/BookLineCustomRepository.java
@@ -8,8 +8,12 @@ import com.floney.floney.book.domain.entity.Book;
 import com.floney.floney.book.domain.entity.BookLine;
 import com.floney.floney.book.domain.entity.BookUser;
 import com.floney.floney.book.domain.entity.Category;
-import com.floney.floney.book.dto.process.*;
+import com.floney.floney.book.dto.process.BookLineExpense;
+import com.floney.floney.book.dto.process.DayLine;
+import com.floney.floney.book.dto.process.DayLineByDayView;
+import com.floney.floney.book.dto.process.TotalExpense;
 import com.floney.floney.book.dto.request.AllOutcomesRequest;
+import com.floney.floney.common.domain.vo.DateDuration;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -18,15 +22,15 @@ import java.util.Optional;
 
 public interface BookLineCustomRepository {
 
-    Map<String, Double> totalExpenseByMonth(String bookKey, DatesDuration dates);
+    Map<String, Double> totalExpenseByMonth(String bookKey, DateDuration dates);
 
-    List<BookLine> findAllBookLineByDuration(String bookKey, DatesDuration duration);
+    List<BookLine> findAllBookLineByDuration(String bookKey, DateDuration duration);
 
     List<DayLineByDayView> allLinesByDay(LocalDate date, String bookKey);
 
     List<TotalExpense> totalExpenseByDay(LocalDate date, String bookKey);
 
-    List<BookLineExpense> dayIncomeAndOutcome(String bookKey, DatesDuration dates);
+    List<BookLineExpense> dayIncomeAndOutcome(String bookKey, DateDuration dates);
 
     void inactiveAllBy(String bookKey);
 
@@ -34,9 +38,9 @@ public interface BookLineCustomRepository {
 
     Double totalExpenseForBeforeMonth(AnalyzeByCategoryRequest request);
 
-    List<AnalyzeResponseByCategory> analyzeByCategory(List<Category> childCategories, DatesDuration duration, String bookKey);
+    List<AnalyzeResponseByCategory> analyzeByCategory(List<Category> childCategories, DateDuration duration, String bookKey);
 
-    Double totalOutcomeMoneyForBudget(AnalyzeRequestByBudget request, DatesDuration duration);
+    Double totalOutcomeMoneyForBudget(AnalyzeRequestByBudget request, DateDuration duration);
 
     Map<String, Double> totalExpensesForAsset(AnalyzeRequestByAsset request);
