+++ src/main/java/com/floney/floney/book/service/BookLineServiceImpl.java
@@ -6,7 +6,6 @@ import com.floney.floney.book.domain.entity.Book;
 import com.floney.floney.book.domain.entity.BookLine;
 import com.floney.floney.book.domain.entity.BookUser;
 import com.floney.floney.book.dto.process.BookLineExpense;
-import com.floney.floney.book.dto.process.DatesDuration;
 import com.floney.floney.book.dto.process.DayLines;
 import com.floney.floney.book.dto.process.TotalExpense;
 import com.floney.floney.book.dto.request.AllOutcomesRequest;
@@ -18,7 +17,7 @@ import com.floney.floney.book.repository.BookLineRepository;
 import com.floney.floney.book.repository.BookRepository;
 import com.floney.floney.book.repository.BookUserRepository;
 import com.floney.floney.book.service.category.CategoryFactory;
-import com.floney.floney.book.util.DateFactory;
+import com.floney.floney.common.domain.vo.DateDuration;
 import com.floney.floney.common.exception.book.NotFoundBookException;
 import com.floney.floney.common.exception.book.NotFoundBookLineException;
 import com.floney.floney.common.exception.book.NotFoundBookUserException;
@@ -66,13 +65,13 @@ public class BookLineServiceImpl implements BookLineService {
     @Override
     public MonthLinesResponse showByMonth(String bookKey, String date) {
         Book book = findBook(bookKey);
-        DatesDuration dates = DateFactory.getStartAndEndOfMonth(date);
+        DateDuration dates = DateDuration.getStartAndEndOfMonth(date);
 
         return MonthLinesResponse.of(
-            date,
-            daysExpense(bookKey, dates),
-            totalExpense(bookKey, dates),
-            carryOverFactory.getCarryOverInfo(book, date)
+                date,
+                daysExpense(bookKey, dates),
+                totalExpense(bookKey, dates),
+                carryOverFactory.getCarryOverInfo(book, date)
         );
     }
 
@@ -84,10 +83,10 @@ public class BookLineServiceImpl implements BookLineService {
         List<TotalExpense> totalExpenses = bookLineRepository.totalExpenseByDay(parse(date), bookKey);
 
         return TotalDayLinesResponse.of(
-            dayLines,
-            totalExpenses,
-            book.getSeeProfile(),
-            carryOverFactory.getCarryOverInfo(book, date)
+                dayLines,
+                totalExpenses,
+                book.getSeeProfile(),
+                carryOverFactory.getCarryOverInfo(book, date)
         );
     }
 
@@ -101,7 +100,7 @@ public class BookLineServiceImpl implements BookLineService {
     @Transactional
     public BookLineResponse changeLine(final BookLineRequest request) {
         final BookLine bookLine = bookLineRepository.findByIdWithCategories(request.getLineId())
-            .orElseThrow(NotFoundBookLineException::new);
+                .orElseThrow(NotFoundBookLineException::new);
         final Book book = findBook(request.getBookKey());
         // TODO: BookLineRequest에 bookKey 삭제 후 아래 메서드 삭제
         validateBookLineIncludedInBook(bookLine.getBook(), book);
@@ -129,7 +128,7 @@ public class BookLineServiceImpl implements BookLineService {
     @Transactional
     public void deleteLine(final Long bookLineId) {
         final BookLine savedBookLine = bookLineRepository.findByIdAndStatus(bookLineId, ACTIVE)
-            .orElseThrow(NotFoundBookLineException::new);
+                .orElseThrow(NotFoundBookLineException::new);
         savedBookLine.inactive();
     }
 
@@ -141,19 +140,19 @@ public class BookLineServiceImpl implements BookLineService {
 
     private BookUser findBookUser(String currentUser, BookLineRequest request) {
         return bookUserRepository.findBookUserByKey(currentUser, request.getBookKey())
-            .orElseThrow(() -> new NotFoundBookUserException(request.getBookKey(), currentUser));
+                .orElseThrow(() -> new NotFoundBookUserException(request.getBookKey(), currentUser));
     }
 
     private Book findBook(String bookKey) {
         return bookRepository.findBookByBookKeyAndStatus(bookKey, ACTIVE)
-            .orElseThrow(() -> new NotFoundBookException(bookKey));
+                .orElseThrow(() -> new NotFoundBookException(bookKey));
     }
 
-    private List<BookLineExpense> daysExpense(String bookKey, DatesDuration dates) {
+    private List<BookLineExpense> daysExpense(String bookKey, DateDuration dates) {
         return bookLineRepository.dayIncomeAndOutcome(bookKey, dates);
     }
 
-    private Map<String, Double> totalExpense(String bookKey, DatesDuration dates) {
+    private Map<String, Double> totalExpense(String bookKey, DateDuration dates) {
         return bookLineRepository.totalExpenseByMonth(bookKey, dates);
     }
 }
